#106 FIX: new tests.
Also the whole suite was reorganized.
