--- samples/AndroidLocationTest.cs
@@ -1,60 +1,0 @@
-﻿using Appium.Samples.Helpers;
-using NUnit.Framework;
-using OpenQA.Selenium;
-using OpenQA.Selenium.Appium;
-using OpenQA.Selenium.Appium.Android;
-using OpenQA.Selenium.Remote;
-using System;
-
-namespace Appium.Samples
-{
-    [TestFixture()]
-    class AndroidLocationTest
-    {
-        private AppiumDriver<AndroidElement> driver;
-        private bool allPassed = true;
-
-        [SetUp]
-        public void BeforeAll()
-        {
-            DesiredCapabilities capabilities = Env.isSauce() ?
-                Caps.getAndroid18Caps(Apps.get("androidApiDemos")) :
-                Caps.getAndroid19Caps(Apps.get("androidApiDemos"));
-            if (Env.isSauce())
-            {
-                capabilities.SetCapability("username", Env.getEnvVar("SAUCE_USERNAME"));
-                capabilities.SetCapability("accessKey", Env.getEnvVar("SAUCE_ACCESS_KEY"));
-                capabilities.SetCapability("name", "android - complex");
-                capabilities.SetCapability("tags", new string[] { "sample" });
-            }
-            Uri serverUri = Env.isSauce() ? AppiumServers.sauceURI : AppiumServers.LocalServiceURIAndroid;
-            driver = new AndroidDriver<AndroidElement>(serverUri, capabilities, Env.INIT_TIMEOUT_SEC);
-            driver.Manage().Timeouts().ImplicitlyWait(Env.IMPLICIT_TIMEOUT_SEC);
-        }
-
-        [TearDown]
-        public void AfterEach()
-        {
-            allPassed = allPassed && (TestContext.CurrentContext.Result.State == TestState.Success);
-            if (Env.isSauce())
-                ((IJavaScriptExecutor)driver).ExecuteScript("sauce:job-result=" + (allPassed ? "passed" : "failed"));
-            driver.Quit();
-            if (!Env.isSauce())
-            {
-                AppiumServers.StopLocalService();
-            }
-        }
-
-        [Test()]
-        public void setLocationTest()
-        {
-            ((AndroidDriver<AndroidElement>)driver).ToggleLocationServices();
-            var l = new Location();
-            l.Altitude = 10;
-            l.Longitude = 10;
-            l.Latitude = 10;
-            driver.Location = l;
-            //var l1 = driver.Location;
-        }
-    }
-}
