+++ samples/iOS/IOSLocationTest.cs
@@ -1,0 +1,68 @@
+﻿using Appium.Samples.Helpers;
+using NUnit.Framework;
+using OpenQA.Selenium;
+using OpenQA.Selenium.Appium;
+using OpenQA.Selenium.Appium.iOS;
+using OpenQA.Selenium.Remote;
+using System;
+
+namespace Appium.Samples.iOS
+{
+    [TestFixture()]
+    class IOSLocationTest
+    {
+        private AppiumDriver<IOSElement> driver;
+        private bool allPassed = true;
+
+        [TestFixtureSetUp]
+        public void beforeAll()
+        {
+            DesiredCapabilities capabilities = Caps.getIos71Caps(Apps.get("iosUICatalogApp"));
+            if (Env.isSauce())
+            {
+                capabilities.SetCapability("username", Env.getEnvVar("SAUCE_USERNAME"));
+                capabilities.SetCapability("accessKey", Env.getEnvVar("SAUCE_ACCESS_KEY"));
+                capabilities.SetCapability("name", "ios - complex");
+                capabilities.SetCapability("tags", new string[] { "sample" });
+            }
+            Uri serverUri = Env.isSauce() ? AppiumServers.sauceURI : AppiumServers.LocalServiceURIForIOS;
+            driver = new IOSDriver<IOSElement>(serverUri, capabilities, Env.INIT_TIMEOUT_SEC);
+            driver.Manage().Timeouts().ImplicitlyWait(Env.IMPLICIT_TIMEOUT_SEC);
+        }
+
+        [TestFixtureTearDown]
+        public void afterAll()
+        {
+            try
+            {
+                if (Env.isSauce())
+                    ((IJavaScriptExecutor)driver).ExecuteScript("sauce:job-result=" + (allPassed ? "passed" : "failed"));
+            }
+            finally
+            {
+                driver.Quit();
+                if (!Env.isSauce())
+                {
+                    AppiumServers.StopLocalService();
+                }
+            }
+        }
+
+        [TearDown]
+        public void AfterEach()
+        {
+            allPassed = allPassed && (TestContext.CurrentContext.Result.State == TestState.Success);
+        }
+
+        [Test()]
+        public void setLocationTest()
+        {
+            var l = new Location();
+            l.Altitude = 10;
+            l.Longitude = 10;
+            l.Latitude = 10;
+            driver.Location = l;
+            //var l1 = driver.Location;
+        }
+    }
+}
